Split out types from chain-events into separate file
