+++ packages/chain-events/src/index.ts
@@ -1,20 +1,12 @@
 export * from './interfaces';
 export * as MolochEvents from './chains/moloch/index';
-export * as MolochTypes from './chains/moloch/types';
 export * as CompoundEvents from './chains/compound/index';
-export * as CompoundTypes from './chains/compound/types';
 export * as SubstrateEvents from './chains/substrate/index';
-export * as SubstrateTypes from './chains/substrate/types';
 export * as Erc20Events from './chains/erc20/index';
-export * as Erc20Types from './chains/erc20/types';
 export * as Erc721Events from './chains/erc721/index';
-export * as Erc721Types from './chains/erc721/types';
 export * as AaveEvents from './chains/aave/index';
-export * as AaveTypes from './chains/aave/types';
 export * as CommonwealthEvents from './chains/commonwealth/index';
-export * as CommonwealthTypes from './chains/commonwealth/types';
 export * as CosmosEvents from './chains/cosmos/index';
-export * as CosmosTypes from './chains/cosmos/types';
 
 export { Listener } from './Listener';
 export * from './handlers';
