+++ packages/commonwealth/client/scripts/controllers/app/webWallets/keplr_web_wallet.ts
@@ -5,7 +5,6 @@ import { OfflineDirectSigner, AccountData } from '@cosmjs/proto-signing';
 
 import { ChainBase, ChainNetwork, WalletId } from 'common-common/src/types';
 import { Account, IWebWallet } from 'models';
-import { validationTokenToSignDoc } from 'adapters/chain/cosmos/keys';
 import { Window as KeplrWindow, ChainInfo } from '@keplr-wallet/types';
 import { StdSignature } from '@cosmjs/amino';
