+++ packages/commonwealth/client/scripts/controllers/chain/ethereum/aave/governance.ts
@@ -1,7 +1,8 @@
 import { ProposalModule, ITXModalData } from 'models';
 import { IApp } from 'state';
 import { IAaveProposalResponse } from 'adapters/chain/aave/types';
-import { AaveEvents, AaveTypes } from 'chain-events/src';
+import { AaveTypes } from 'chain-events/src/types';
+import { AaveEvents } from 'chain-events/src';
 import { Executor } from 'common-common/src/eth/types';
 import { chainToEventNetwork } from 'controllers/server/chain_entities';
