+++ packages/commonwealth/client/scripts/controllers/chain/ethereum/aave/proposal.ts
@@ -3,7 +3,7 @@ import BN from 'bn.js';
 import Web3 from 'web3';
 import $ from 'jquery';
 import bs58 from 'bs58';
-import { AaveTypes } from 'chain-events/src';
+import { AaveTypes } from 'chain-events/src/types';
 import { ProposalType } from 'common-common/src/types';
 import { EthereumCoin } from 'adapters/chain/ethereum/types';
 import { IAaveProposalResponse } from 'adapters/chain/aave/types';
