+++ packages/commonwealth/client/scripts/controllers/chain/substrate/bountyTreasury.ts
@@ -6,7 +6,7 @@ import {
   SubstrateCoin
 } from 'adapters/chain/substrate/types';
 import { ProposalModule } from 'models';
-import { SubstrateTypes } from 'chain-events/src';
+import { SubstrateTypes } from 'chain-events/src/types';
 import SubstrateChain from './shared';
 import SubstrateAccounts, { SubstrateAccount } from './account';
 import { SubstrateBounty } from './bounty';
