+++ packages/commonwealth/client/scripts/controllers/chain/substrate/collective_proposal.ts
@@ -1,6 +1,6 @@
 import _ from 'underscore';
 
-import { SubstrateTypes } from 'chain-events/src';
+import { SubstrateTypes } from 'chain-events/src/types';
 import { ApiPromise } from '@polkadot/api';
 import { Votes } from '@polkadot/types/interfaces';
 import { Option } from '@polkadot/types';
