+++ packages/commonwealth/client/scripts/controllers/chain/substrate/democracy_referendum.ts
@@ -19,7 +19,7 @@ import {
   ChainEntity,
   ChainEvent,
 } from 'models';
-import { SubstrateTypes } from 'chain-events/src';
+import { SubstrateTypes } from 'chain-events/src/types';
 import { Coin } from 'adapters/currency';
 import SubstrateChain from './shared';
 import SubstrateAccounts, { SubstrateAccount } from './account';
