+++ packages/commonwealth/client/scripts/controllers/chain/substrate/treasury_tips.ts
@@ -2,7 +2,7 @@ import { ApiPromise } from '@polkadot/api';
 import { AccountId } from '@polkadot/types/interfaces';
 import { Vec } from '@polkadot/types';
 import { ISubstrateTreasuryTip, SubstrateCoin } from 'adapters/chain/substrate/types';
-import { SubstrateTypes } from 'chain-events/src';
+import { SubstrateTypes } from 'chain-events/src/types';
 import { ProposalModule } from 'models';
 import { IApp } from 'state';
 import { formatAddressShort } from 'utils';
