+++ packages/commonwealth/client/scripts/views/components/proposal_card/helpers.tsx
@@ -3,7 +3,7 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import m from 'mithril'; // required for getStatusText
 import moment from 'moment';
-import { AaveTypes, CompoundTypes } from 'chain-events/src';
+import { AaveTypes, CompoundTypes } from 'chain-events/src/types';
 
 import 'components/proposal_card/index.scss';
