+++ packages/commonwealth/client/scripts/views/components/proposals/helpers.ts
@@ -16,7 +16,7 @@ import CompoundProposal, {
   CompoundProposalVote,
 } from 'controllers/chain/ethereum/compound/proposal';
 import { notifyError } from 'controllers/app/notifications';
-import { CompoundTypes } from 'chain-events/src';
+import { CompoundTypes } from 'chain-events/src/types';
 import NearSputnikProposal from 'controllers/chain/near/sputnik/proposal';
 import {
   NearSputnikProposalStatus,
