+++ packages/commonwealth/server/util/proposalIdToEntity.ts
@@ -1,5 +1,5 @@
 import { ProposalType } from 'common-common/src/types';
-import { SubstrateTypes, MolochTypes, CompoundTypes, AaveTypes } from 'chain-events/src';
+import { SubstrateTypes, MolochTypes, CompoundTypes, AaveTypes } from 'chain-events/src/types';
 
 // this function takes an "old style" identifier such as treasuryproposal_4 and attempts
 // fetch the corresponding chain entity from the database
