+++ packages/commonwealth/webpack/webpack.config.dev.js
@@ -20,7 +20,7 @@ module.exports = merge(common, {
   },
   mode: 'development',
   target: 'web',
-  devtool: 'eval-cheap-source-map',
+  devtool: 'source-map',
   output: {
     publicPath: '/build',
     path: path.join(__dirname, '../build'),
